refactor(editor): toolbar floating layer 분리
