+++ front/e2e/editor-studio-state.spec.ts
@@ -331,6 +331,14 @@ test.describe("editor studio state", () => {
     const blockEditorEngineStylesSource = existsSync(blockEditorEngineStylesPath)
       ? readFileSync(blockEditorEngineStylesPath, "utf8")
       : ""
+    const blockEditorEngineLayersPath = path.resolve(
+      __dirname,
+      "../src/components/editor/BlockEditorEngine.layers.tsx"
+    )
+    expect(existsSync(blockEditorEngineLayersPath)).toBe(true)
+    const blockEditorEngineLayersSource = existsSync(blockEditorEngineLayersPath)
+      ? readFileSync(blockEditorEngineLayersPath, "utf8")
+      : ""
     const blockSelectionModelSource = readFileSync(
       path.resolve(__dirname, "../src/components/editor/blockSelectionModel.ts"),
       "utf8"
@@ -854,8 +862,23 @@ test.describe("editor studio state", () => {
     expect(blockEditorShellSource).toContain('import BlockEditorEngine from "./BlockEditorEngine"')
     expect(blockEditorShellSource).toContain('import type { BlockEditorEngineProps } from "./blockEditorEngineTypes"')
     expect(blockEditorShellSource).toContain("const BlockEditorShell = (props: BlockEditorEngineProps) => <BlockEditorEngine {...props} />")
-    expect(blockEditorEngineSource.split("\n").length).toBeLessThanOrEqual(8500)
+    expect(blockEditorEngineSource.split("\n").length).toBeLessThanOrEqual(7000)
     expect(blockEditorEngineSource).toContain('} from "./BlockEditorEngine.styles"')
+    expect(blockEditorEngineSource).toContain('} from "./BlockEditorEngine.layers"')
+    expect(blockEditorEngineLayersSource).toContain("export const BlockEditorToolbarLayer =")
+    expect(blockEditorEngineLayersSource).toContain("export const BlockEditorQuickInsertLayer =")
+    expect(blockEditorEngineLayersSource).toContain("export const BlockEditorSlashMenuLayer =")
+    expect(blockEditorEngineLayersSource).toContain("export const BlockEditorFloatingBubbleLayer =")
+    expect(blockEditorEngineLayersSource).toContain("export const BlockEditorBlockHandleLayer =")
+    expect(blockEditorEngineLayersSource).toContain('data-testid="slash-menu"')
+    expect(blockEditorEngineLayersSource).toContain('data-testid="editor-text-bubble-toolbar"')
+    expect(blockEditorEngineLayersSource).toContain('"block-drag-handle"')
+    expect(blockEditorEngineSource).not.toContain("<ToolbarActions>")
+    expect(blockEditorEngineSource).not.toContain("<QuickInsertBar")
+    expect(blockEditorEngineSource).not.toContain("<SlashMenu")
+    expect(blockEditorEngineSource).not.toContain("<FloatingBubbleToolbar")
+    expect(blockEditorEngineSource).not.toContain("<BlockHandleRail")
+    expect(blockEditorEngineSource).not.toContain("<FloatingBlockMenu")
     expect(blockEditorEngineSource).not.toContain("const Shell = styled.div`")
     expect(blockEditorEngineSource).not.toContain("const Toolbar = styled.div`")
     expect(blockEditorEngineSource).not.toContain("const EditorViewport = styled.div`")
@@ -1239,6 +1262,10 @@ test.describe("editor studio state", () => {
       path.resolve(__dirname, "../src/components/editor/BlockEditorEngine.styles.tsx"),
       "utf8"
     )
+    const blockEditorEngineLayersSource = readFileSync(
+      path.resolve(__dirname, "../src/components/editor/BlockEditorEngine.layers.tsx"),
+      "utf8"
+    )
     const tableAffordanceModelSource = readFileSync(
       path.resolve(__dirname, "../src/components/editor/tableAffordanceModel.ts"),
       "utf8"
@@ -1304,8 +1331,8 @@ test.describe("editor studio state", () => {
       "utf8"
     )
 
-    expect(blockEditorEngineSource).toContain('data-testid="table-column-drag-guide"')
-    expect(blockEditorEngineSource).toContain('data-testid={`table-column-resize-boundary-${index}`}')
+    expect(blockEditorEngineLayersSource).toContain('data-testid="table-column-drag-guide"')
+    expect(blockEditorEngineLayersSource).toContain('data-testid={`table-column-resize-boundary-${index}`}')
     expect(blockEditorEngineSource).toContain('} from "./tableResizeInteractionModel"')
     expect(tableResizeInteractionModelSource).toContain("export type TableRowResizeState = {")
     expect(tableResizeInteractionModelSource).toContain("export type TableColumnRailResizeState = {")
@@ -1344,19 +1371,19 @@ test.describe("editor studio state", () => {
     expect(blockEditorEngineSource).not.toContain("const startTableColumnResizeFromDomHandle = useCallback(")
     expect(blockEditorEngineSource).toContain('} from "./tableFloatingUiModel"')
     expect(blockEditorEngineSource).toContain("const getActiveTableRectFromDom = useCallback(")
-    expect(blockEditorEngineSource).toContain('import { createPortal } from "react-dom"')
-    expect(blockEditorEngineSource).toContain("const tableOverlayPortal =")
-    expect(blockEditorEngineSource).toContain("createPortal(tableOverlay, document.body)")
+    expect(blockEditorEngineLayersSource).toContain('import { createPortal } from "react-dom"')
+    expect(blockEditorEngineLayersSource).toContain("const tableOverlay = (")
+    expect(blockEditorEngineLayersSource).toContain("createPortal(tableOverlay, document.body)")
     expect(blockEditorEngineSource).toContain("const TABLE_EDGE_HANDLE_INSET_PX = 6")
     expect(tableAffordanceModelSource).toContain("export const TABLE_EDGE_ADD_BUTTON_SIZE_PX = 24")
     expect(tableAffordanceModelSource).toContain("export const TABLE_CELL_MENU_BUTTON_SIZE_PX = 22")
     expect(tableAffordanceModelSource).toContain("export const TABLE_ADD_BAR_VIEWPORT_PADDING_PX = 8")
-    expect(blockEditorEngineSource).toContain('data-testid="table-corner-grow-handle"')
-    expect(blockEditorEngineSource).toContain('data-testid="table-corner-preview-outline"')
-    expect(blockEditorEngineSource).toContain('data-testid="table-structure-menu-button"')
-    expect(blockEditorEngineSource).toContain('data-testid="table-cell-menu-button"')
-    expect(blockEditorEngineSource).toContain('data-testid="table-overflow-mode-normal"')
-    expect(blockEditorEngineSource).toContain('data-testid="table-overflow-mode-wide"')
+    expect(blockEditorEngineLayersSource).toContain('data-testid="table-corner-grow-handle"')
+    expect(blockEditorEngineLayersSource).toContain('data-testid="table-corner-preview-outline"')
+    expect(blockEditorEngineLayersSource).toContain('data-testid="table-structure-menu-button"')
+    expect(blockEditorEngineLayersSource).toContain('data-testid="table-cell-menu-button"')
+    expect(blockEditorEngineLayersSource).toContain('data-testid="table-overflow-mode-normal"')
+    expect(blockEditorEngineLayersSource).toContain('data-testid="table-overflow-mode-wide"')
     expect(tableWidthModelSource).toContain('export const TABLE_OVERFLOW_MODE_WIDE = "wide"')
     expect(tableWidthModelSource).toContain("export const getTableOverflowMode = (")
     expect(tableWidthModelSource).toContain("export const shouldPromoteWideTableOverflowMode = (")
@@ -1419,9 +1446,9 @@ test.describe("editor studio state", () => {
     expect(blockEditorEngineSource).not.toContain("const resolveTableScopedSelectedCell = (")
     expect(blockEditorEngineSource).not.toContain("const resolveActiveRenderedTableForFloatingUi = (")
     expect(blockEditorEngineSource).not.toContain("const readRenderedColumnWidths = (")
-    expect(blockEditorEngineSource).toContain('data-testid="table-row-drag-shadow"')
-    expect(blockEditorEngineSource).toContain('"table-row-reorder-indicator"')
-    expect(blockEditorEngineSource).toContain('"table-column-reorder-indicator"')
+    expect(blockEditorEngineLayersSource).toContain('data-testid="table-row-drag-shadow"')
+    expect(blockEditorEngineLayersSource).toContain('"table-row-reorder-indicator"')
+    expect(blockEditorEngineLayersSource).toContain('"table-column-reorder-indicator"')
     expect(blockEditorEngineStylesSource).toContain('const TableCellMenuButton = styled(TableHandleButton)`')
     expect(blockEditorEngineSource).toContain("const updateActiveTableOverflowMode = useCallback(")
     expect(blockEditorEngineSource).toContain("const reorderTableAxisAtPosition = useCallback(")
@@ -1452,16 +1479,16 @@ test.describe("editor studio state", () => {
     expect(markdownRendererRootSource).toContain("max-width: none;")
     expect(blockEditorEngineSource).toContain("isRowMenuOpen,")
     expect(blockEditorEngineSource).toContain("isColumnMenuOpen,")
-    expect(blockEditorEngineSource).toContain("activeTableStructureState.hasHeaderRow")
-    expect(blockEditorEngineSource).toContain("activeTableStructureState.hasHeaderColumn")
+    expect(blockEditorEngineLayersSource).toContain("activeTableStructureState.hasHeaderRow")
+    expect(blockEditorEngineLayersSource).toContain("activeTableStructureState.hasHeaderColumn")
     expect(blockEditorEngineSource).toContain("const tableCornerGrowStepMetrics = resolveTableCornerGrowStepMetrics(tableAffordanceGeometry)")
-    expect(blockEditorEngineSource).toContain("data-column-step={tableCornerGrowStepMetrics.columnStepPx}")
-    expect(blockEditorEngineSource).toContain("data-row-step={tableCornerGrowStepMetrics.rowStepPx}")
-    expect(blockEditorEngineSource).toContain('aria-label="표 구조 메뉴"')
-    expect(blockEditorEngineSource).toContain("페이지 너비에 맞춤")
-    expect(blockEditorEngineSource).toContain("넓은 표")
-    expect(blockEditorEngineSource).toContain("제목 행")
-    expect(blockEditorEngineSource).toContain("제목 열")
+    expect(blockEditorEngineLayersSource).toContain("data-column-step={tableCornerGrowStepMetrics.columnStepPx}")
+    expect(blockEditorEngineLayersSource).toContain("data-row-step={tableCornerGrowStepMetrics.rowStepPx}")
+    expect(blockEditorEngineLayersSource).toContain('aria-label="표 구조 메뉴"')
+    expect(blockEditorEngineLayersSource).toContain("페이지 너비에 맞춤")
+    expect(blockEditorEngineLayersSource).toContain("넓은 표")
+    expect(blockEditorEngineLayersSource).toContain("제목 행")
+    expect(blockEditorEngineLayersSource).toContain("제목 열")
     expect(tableAffordanceModelSource).toContain("export type TableAffordanceGeometry = {")
     expect(tableAffordanceModelSource).toContain("export type TableAffordanceVisibility = {")
     expect(tableAffordanceModelSource).toContain("export const INITIAL_TABLE_AFFORDANCE_GEOMETRY: TableAffordanceGeometry = {")
